Fixed a bug

+++ Engine/Player.h
@@ -45,6 +45,6 @@ public:
 	void Draw(Graphics& gfx);
 	void Update(Keyboard& kbd, Mouse& micky, const float dt);
 	Vec2 GetPosition() const;
-	int GetCameraX() const;
-	int GetCameraY() const;
+	float GetCameraX() const;
+	float GetCameraY() const;
 };
